Wiktionary Index creation now uses Stax2 so that location info is encoded in a long, allowing for the parsing of >2Gb XML files Offset information updated to use a long offset for article start Wiktionary index dump now stores long offsets Wiktionary index signature updated to avoid reuse of old index format

+++ modules/wiktionary-support/src/main/java/org/getalp/blexisma/wiktionary/OffsetValue.java
@@ -7,10 +7,10 @@ public class OffsetValue implements Serializable {
      * 
      */
     private static final long serialVersionUID = -570466055812338894L;
-    public int start;
+    public long start;
 	public int length;
 
-	public OffsetValue(int start, int length) {
+	public OffsetValue(long start, int length) {
 		this.start = start;
 		this.length = length;
 	}
